Now works with up to date packages
The main problem was react router has been updated to V6. You now have
to arrange the router code a little differently.

I have completely rewritten the protected route HOC, it's now just a
wrapper component that either returns it's child component (the
protected route) or redirects you to the login screen.

The bug where you catch a glimpse of the protected route after the JWT
has expired and you click on it is still prevalent!