Restructured App.js
I have separated the logic from the JSX in the main App Component App.js
now consists of App.js and AppView.js. It is a little different from the
other components because it is a class component.

Next I'm going to transfer App.js's state to a Redux store...