Delete user working
also there are now properly handled errors concerning the table data
fetch in <Body />

The dbConfig.js file now has an extra option for a table prefix, I have
also made the values in the config file referance dbconfig.json which
is to be used to set the project up during deployment. All sql queries
in the server code now get the database name and the table name
(including the prefix if there is one) from the db.connection object.

I am in the process of making a bash script that will generate
dbconfig.json according to user input and then create the database and
user table. It doesn't work yet tho (see/server/configuredb.sh) and it's
the weekend now!