Fixed numerous issues
Login server errors are now handled properly and sent back to the client

I have standardised all the error messages to use the same key names to
improve interoperability

The table cells now are now formatted to overflow: elipsis with with
scrollable content and no scrollbars.

The custom style object passed to the dynamicTable now only has maxWidth
in and it works!

Other style changes to the dynamicTable

The login and sign-up buttons are now disabled when logged in

Other style changes to the header