Login email and pwd no sent base64 auth header
Using btoa() and atob() to encode auth header

removed request body as no longer needed

there's a bit of jiggery-pokery decoding the password because I'm using
a ':' character to separate email and password which means it would get
stripped out if I rely soley on the String.split(':') function, I.e you
wouldn't be able to use a ':' in your password. Fixed it with a regExp
in the split() function and a small loop to reintegrate split ':' chars.

also fixed a but where the login details would remain in the login form
fields after the token has expired. Fixed by clearing fields after login